Add .gitattributes for Unity's files (#2)
Marks Unity-generated files as generated so that their diff is hidden by
default in GitHub's UI. You can still use "Load diff" to load it, it's
just that it doesn't load of all of these diffs slowing down the whole
thing. Additionally, I've added the `merge=unityyamlmerge` attribute to
those files to automatically use the
[UnityYAMLMerge](https://docs.unity3d.com/Manual/SmartMerge.html) when
conflict resolution of these files is required.

We may want to also consider using Git LFS but it does slightly affect
the developer's workflow so I opted to not touch that until we decide to
use that.