Make Point instances mutable (#13)
This PR more or less resolves the concern about it being hard to modify
a point shared between polygons. `SolidFigure` currently makes no effort
to figure out if a point with the same coordinates but different
references are shared between `Polygon` instances *but* if multiple
`Polygon`s receive the same reference to a `Point` then any modification
of that `Point` will affect all of them (and all relevant
`PolygonPresenter`s will update the meshes appropriately as well).