Implement rendering of solids of revolution (+ refactoring) (#15)
Implements models and rendering of solids of revolution that are of
interest to us (sphere, cylinder, cone, truncated cone) and refactors
out existing implementations for polyhedrons (which were previously just
called solid figures). I added a way to render a basic circle as well
which is currently used for bases in cylinders/cones and for an equator
of the sphere.

The meshes are generated using UV mapping, the implementation is based
on what I found in this project https://github.com/kurtdekker/makegeo
though there are similar implementations available here as well:
-
https://web.archive.org/web/20210828182001/https://wiki.unity3d.com/index.php/ProceduralPrimitives#C.23_-_Cone
-
https://web.archive.org/web/20210828182001/https://wiki.unity3d.com/index.php/ProceduralPrimitives#C.23_-_Sphere

There are more methods for generating meshes for e.g. a sphere than the
UV mapping one I used but it is probably the simplest available. Here's
an interesting comparison of the accuracy of these methods (dependent on
the method and a number of subdivisions):

https://medium.com/@oscarsc/four-ways-to-create-a-mesh-for-a-sphere-d7956b825db4

Some other methods, also not used:
- https://danielsieger.com/blog/2021/03/27/generating-spheres.html